+++ package.json
@@ -9,6 +9,7 @@
     "firebase": "^9.6.10",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
